Upgrades to Arrow 2.x (#83)
* Upgrades arrow to 2.0.0-alpha.2

* Removes tests for Validated

Validated has been deprecated, and is no longer included in Arrow.

* Backports ValidatedNel as Either<NonEmptyList<E>, A>

This patch backports ValidatedNel as an alias of an underlying type, and
maintains the existing extensions.

* Backports Option's toValidatedNel to use Either<NonEmptyList<E>, A>

* Uses the Raise DSL instead of continuations

Continuations have been deprecated in favour of the Raise DSL.

I have removed OutcomeEagerEffect scope, as the Raise DSL does not
expose an eager implementation to be extend. Instead, the expectation is
that the caller wraps calls in `eagerEffect`. This binds the suspended
function to an eager context.

* Removes unused import

* Re-adds OutcomeEagerEffectScope to provide a suspended and blocking bind

* Adds arbNel in place of Arb.nonEmptyList

kotest-extensions-arrow has not yet been upgrade to arrow 2.0.0, so
we're seeing compatibility issues with some of the older
implementations.

In the meantime, we can add our own Arb to make tests pass.

* Updates API declarations

* Removes OutcomeEffectScope and OutcomeEagerEffectScope

`OutcomeBuilder` should be used instead, with the following notation:
```kt
outcome {
   Present(1).bind()
}
```

* Uses Arb.nel from io.kotest.extensions:kotest-property-arrow

* Updates arrow, kotest, and kotlin to latest

* Updates kotlinBinaryCompatibilityPlugin

* Updates API declarations

* Removes unused imports

* Updates documentation for outcome

* Updates changelog